Added error handling to controller's methods

+++ product-recommendation-service/src/main/java/com/rainbowforest/recommendationservice/controller/RecommendationController.java
@@ -2,18 +2,16 @@ package com.rainbowforest.recommendationservice.controller;
 
 import com.rainbowforest.recommendationservice.feignClient.ProductClient;
 import com.rainbowforest.recommendationservice.feignClient.UserClient;
+import com.rainbowforest.recommendationservice.http.header.HeaderGenerator;
 import com.rainbowforest.recommendationservice.model.Product;
 import com.rainbowforest.recommendationservice.model.Recommendation;
 import com.rainbowforest.recommendationservice.model.User;
 import com.rainbowforest.recommendationservice.service.RecommendationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.List;
 
 @RestController
@@ -27,34 +25,75 @@ public class RecommendationController {
 
     @Autowired
     private UserClient userClient;
+    
+    @Autowired
+    private HeaderGenerator headerGenerator;
 
     @GetMapping(value = "/recommendations")
-    private List<Recommendation> getAllRating(@RequestParam("name") String productName){
-        return recommendationService.getAllRecommendationByProductName(productName);
+    private ResponseEntity<List<Recommendation>> getAllRating(@RequestParam("name") String productName){
+        List<Recommendation> recommendations = recommendationService.getAllRecommendationByProductName(productName);
+        if(!recommendations.isEmpty()) {
+        	return new ResponseEntity<List<Recommendation>>(
+        		recommendations,
+        		headerGenerator.getHeadersForSuccessGetMethod(),
+        		HttpStatus.OK);
+        }
+        return new ResponseEntity<List<Recommendation>>(
+        		headerGenerator.getHeadersForError(),
+        		HttpStatus.NOT_FOUND);
     }
-
+    
     @PostMapping(value = "/{userId}/recommendations/{productId}")
     private ResponseEntity<Recommendation> saveRecommendations(
             @PathVariable ("userId") Long userId,
             @PathVariable ("productId") Long productId,
             @RequestParam ("rating") int rating,
-            HttpServletRequest request) throws URISyntaxException {
-
-        HttpHeaders headers = new HttpHeaders();
-        Recommendation recommendation = new Recommendation();
-        Product product = productClient.getProductById(productId);
-        User user = userClient.getOneUser(userId);
-        recommendation.setProduct(product);
-        recommendation.setUser(user);
-        recommendation.setRating(rating);
-        recommendationService.saveRecommendation(recommendation);
-        headers.setLocation(new URI(request.getRequestURI() + "/" + recommendation.getId()));
-        return new ResponseEntity<Recommendation>(recommendation, headers, HttpStatus.CREATED);
+            HttpServletRequest request){
+    	
+    	Product product = productClient.getProductById(productId);
+		User user = userClient.getUserById(userId);
+    	
+		if(product != null && user != null) {
+			try {
+				Recommendation recommendation = new Recommendation();
+				recommendation.setProduct(product);
+				recommendation.setUser(user);
+				recommendation.setRating(rating);
+				recommendationService.saveRecommendation(recommendation);
+				return new ResponseEntity<Recommendation>(
+						recommendation,
+						headerGenerator.getHeadersForSuccessPostMethod(request, recommendation.getId()),
+						HttpStatus.CREATED);
+			}catch (Exception e) {
+				e.printStackTrace();
+				return new ResponseEntity<Recommendation>(
+						headerGenerator.getHeadersForError(),
+						HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		}
+        return new ResponseEntity<Recommendation>(
+        		headerGenerator.getHeadersForError(),
+        		HttpStatus.BAD_REQUEST);
     }
 
     @DeleteMapping(value = "/recommendations/{id}")
-    private ResponseEntity deleteRecommendations(@PathVariable("id") Long id){
-        recommendationService.deleteRecommendation(id);
-        return new ResponseEntity(HttpStatus.OK);
+    private ResponseEntity<Void> deleteRecommendations(@PathVariable("id") Long id){
+    	Recommendation recommendation = recommendationService.getRecommendationById(id);
+    	if(recommendation != null) {
+    		try {
+    			recommendationService.deleteRecommendation(id);
+    			return new ResponseEntity<Void>(
+    					headerGenerator.getHeadersForSuccessGetMethod(),
+    					HttpStatus.OK);
+    		}catch (Exception e) {
+    			e.printStackTrace();
+    			return new ResponseEntity<Void>(
+    					headerGenerator.getHeadersForError(),
+    					HttpStatus.INTERNAL_SERVER_ERROR);	
+    		}
+    	}
+    	return new ResponseEntity<Void>(
+    			headerGenerator.getHeadersForError(),
+    			HttpStatus.NOT_FOUND);
     }
 }
